[Chore] Separate out system utilities from vllm.utils (#27201)
